# src/Entity/Shop.php
<?php

declare(strict_types=1);

namespace App\Entity;

use Doctrine\ORM\Mapping as ORM;
use OpenApi\Attributes as OA;

#[ORM\Entity]
#[ORM\Table(name: 'shops')]
class Shop
{
    #[ORM\Id]
    #[ORM\GeneratedValue(strategy: 'AUTO')]
    #[ORM\Column(type: 'integer')]
    private int $id;

    #[ORM\Column(type: 'string', length: 255)]
    #[OA\Property(property: 'title', type: 'string')]
    private string $title;
    #[ORM\Column(type: 'string', length: 255)]
    #[OA\Property(property: 'address', type: 'int')]
    private string $address;
    #[ORM\Column(type: 'float')]
    private float $latitude;
    #[ORM\Column(type: 'float')]
    private float $longitude;

    #[ORM\Column(type: 'string', length: 255)]
    private string $description;

    public function __construct(
        string $title,
        string $address,
        float $latitude,
        float $longitude,
        string $description,
    ) {
        $this->title = $title;
        $this->address = $address;
        $this->latitude = $latitude;
        $this->longitude = $longitude;
        $this->description = $description;
    }

    public function getTitle(): string
    {
        return $this->title;
    }

    public function setTitle(string $title): void
    {
        $this->title = $title;
    }

    public function getAddress(): string
    {
        return $this->address;
    }

    public function setAddress(string $address): void
    {
        $this->address = $address;
    }

    public function getLatitude(): float
    {
        return $this->latitude;
    }

    public function setLatitude(float $latitude): void
    {
        $this->latitude = $latitude;
    }

    public function getLongitude(): float
    {
        return $this->longitude;
    }

    public function setLongitude(float $longitude): void
    {
        $this->longitude = $longitude;
    }

    public function getDescription(): string
    {
        return $this->description;
    }
}